taskbolt/newx/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
PyYAML>=6.0

celery[redis]>=5.3.0

langgraph>=0.0.30

GitPython>=3.1.0
